external_work: Add .call_in_thread() and .run_in_subprocess() to Job API
This moves the API for scheduling work in another thread/process from
the Scheduler class and into the Job class. This allows Job subclasses
to customize how this work scheduling is done. For now, it allows jobs
that use the LogMux class to automatically set up its subprocesses to
write to the logmuxed stdout/stderr streams.