basic.Scheduler: Fix failure -> cancellation of dependent jobs in Python >=v3.8
Starting with Python v3.8, asyncio.CancelledError no longer inherits
from concurrent.futures.CancelledError, hence when we raise the latter
in Scheduler.results() (as a consequence of failure in a job we were
awaiting), the task would be counted as "failed" rather than
"cancelled". AFAICS, there's no reason why we should not simply raise
the former instead.