Add polyfill module w/async-related niceness from Python >v3.6
There are several nice async-related additions in Python v3.7 and v3.8
that we'd like to use, even though we must remain compatible with v3.6.
The polyfill module patches up the contextlib and asyncio modules to
contain the functionality we want from the newer versions.

This allows us to write code elsewhere in the package as if we're always
running on Python v3.8.