basic.Scheduler: Introduce .add_job() to move focus away from Job
We want to decouple the Job class from the Scheduler class and move
towards elimintaing the Job class entirely. This pathch introduces a
method for adding job coroutines to the Scheduler without them having to
be part of a Job instance. Instead, the name, coroutine, and
dependencies are passed as separate arguments.