logmuxed_work: Integrate log_context.LogContextDemuxer
This is a major revamp in how the logmuxed_work scheduler interacts with
logging: Before this patch, a logmuxed job (upon entering the redirected
context) would simply add a logging.StreamHandler to the Context.logger
instance in order to have those log records written to the redirected
stderr. Logging calls made from other loggers would bypass the logmux
completely.

Now, we start using the LogContextDemuxer instead, to setup log handlers
that redirect _all_ log records based on which job they come from:

  - Upon entering the `redirect` context, we register a context handler
    for the current asyncio task to redirect logs into the job's FIFO.
    This assumes there is a one-to-one correspondence between jobs and
    tasks (at least for the duration of the `redirect` context). If the
    job were to start processing in another task, logs from there would
    not be redirected.

  - Upon starting work in a worker thread .call_in_thread() will register
    a context handler for that worker thread, so that all logs from that
    thread are redirected into the job's FIFO.

  - For subprocesses, we're already good, as we already redirect their
    stdout/stderr into the job's FIFO, and any logs made from the
    subprocess will (typically) go to its stderr.

The LogContextDemuxer object is managed by the logmuxed_work.Scheduler,
and we automatically install it into the root logger for the duration of
Scheduler._run_tasks().

Note that this means there is no more need for a separate logger instance
per job (i.e. Context.logger - will be removed in the next commit).
Instead, now _all_ log messages should be redirected properly from within
a redirected context.

The tests are significantly updated/expanded to incorporate the testing
of log records ending up in the right place in the various scenarios.