Remove Context.logger, use module-level loggers instead.
We don't want to create a logger instance for every single job because
of the reasons outlined in [1] as well as other reasons outlined in the
commit introducing the logcontext module.

Now that the logcontext is integrated, we can finally remove the per-job
logger.

Adjust test and example code accordingly.

[1]: Quoting the "Adding contextual information to your logging output"
section on https://docs.python.org/3/howto/logging-cookbook.html:

  While it might be tempting to create Logger instances on a per-[job]
  basis, this is not a good idea because these instances are not garbage
  collected. While this is not a problem in practice, when the number of
  Logger instances is dependent on the level of granularity you want to
  use in logging an application, it could be hard to manage if the
  number of Logger instances becomes effectively unbounded.