conftest: Scale stress tests back a bit to lower test run times
The stress tests in test_3_abort.py and test_5_logmuxed_work.py run 100
jobs across scheduler instances with 1, 2, 4 and 100 workers. This takes
up the bulk of the test suite run time, which is rapidly approaching a
whole minute. Especially some of the tests run very slowly when the
number of workers is 1 or 2, and running 100 jobs in (almost-)series
does not add much value over running 10 jobs in the same scenario.

Therefore, scale back the number of jobs run in these stress tests, from
a static 100 to a more dynamic min(100, 10 * workers). That is, each
stress test will run:

 - 1 worker: 10 jobs
 - 2 workers: 20 jobs
 - 4 workers: 40 jobs
 - 100 workers: 100 jobs

This puts the total test suite run time back into the ~30s range.