external_work: Don't attempt exception recovery on subprocess termination
When we time out waiting for a subprocess to be terminated,
asyncio.wait_for() will cancel the proc.wait() and wait for the
cancellation to complete/propagate. However, starting with Python v3.9,
(as part of the fix for https://bugs.python.org/issue40607) the
CancelledError that is thrown into the proc.wait() and propagates back
out is itself stored as the e.__context__ of the TimeoutError raised by
asyncio.wait_for().

This interferes with our misguided attempt to preserve/propagate
e.__context__ from that TimeoutError as a somehow "true" reason for why
the process failed, when in fact, the impetus for its failing was quite
obviously the SIGTERM/SIGKILL we just sent it.

I suspect the real reason for the e.__context__ logic was to avoid
propagating any exception at all from the TimeoutError which is what
happens on Python <=v3.8 where e.__context__ is (always?) None. Make
sure this now also happens on >v3.8.

Making sure a terminated/killed subprocess ends up propagating an
exception back to the surrounding job code is in any case accomplished
by using check=True, which will cause CalledProcessError to be raised
due to the non-zero exit code (-15 for SIGTERM or -9 for SIGKILL).