test_external_work: Fix asyncio.gather leaving unfinished tasks around
We see a couple of hard-to-debug test failures on Travis/Github Actions,
where an earlier test has left behind an orphan task, which at a later
point completes and reports an ignored exception. This then shows up in
the captured output of a later test, which causes that to fail.

The earlier test(s) in this case are tests that use asyncio.gather() to
schedule two concurrent threads/subprocess from one job. When the number
of workers is 1, this (expectedly) raises RuntimeError when trying to
instantiate the second thread/subprocess. The RuntimeError is
immediately propagated by asyncio.gather() _without_ waiting for the
_first_ thread/subprocess to complete. It is instead left running in the
background, and messes up a later test.

This is in fact documented by asyncio.gather()[1]:

  If 'return_exceptions' is False (default), the first raised exception
  is immediately propagated to the task that awaits on gather(). Other
  awaitables in the 'aws' sequence _won’t be cancelled_ and will
  continue to run.

Fix this by passing return_exceptions=True to asyncio.gather() so that
we will always wait until all awaitables are complete. Then iterate
through the results and raise any exceptions found.

This has a positive effect on the events emitted, as we'll now always
see the "finish work in subprocess" event for the first subprocess, and
we can drop the somewhat awkward 'abort' handling in TExternalWorkJob's
.subprocess_xevents() helper.

[1]: https://docs.python.org/3/library/asyncio-task.html#asyncio.gather