logmux: Remove lambda requirement for simple decorators
Most decorator simply prepend a fixed prefix (or append a fixed suffix)
to each line in the logmux stream. These uses currently pass need to
pass something like this:

    async with logmux.new_stream(
        lambda line: f"[pre] {line.rstrip()} [post]\n"
    ) as f:
        ...

This is complicated and error-prone for a very common use case.

Make usage easier by providing a logmux.simple_decorator() helper
function that takes a prefix and (optional) suffix, and returns the
corresponding lambda that users previously had to write themselves. Also
teach LogMux to recognize when a (prefix) string is passed directly as a
decorator and auto-wrap it in logmux.simple_decorator().

This allows all prefix + suffix decorator lambdas to be replaced by
calls to logmux.simple_decorator() instead:

    async with logmux.new_stream(
        logmux.simple_decorator('[pre] ', ' [post]')
    ) as f:
        ...

Furthermore, all prefix decorators can be simplified to simply passing
the prefix string directly:

    async with logmux.new_stream('[pre] ') as f:
        ...