logmux: Work on binary streams instead of text streams
Given that we have ultimately no control of the output of users' threads
and subprocesses (especially when e.g. using asyncjobs to run 3rd-party
code in subprocesses), it is risky to handle the logmux FIFO as UTF-8-
encoded text streams (even if we do use the 'surrogateescape' handler to
preserve non-decodable bytes).

Open the logmux FIFOS in binary mode, and write a binary output stream
instead. This does mean that all decorators now have to be bytes-based
and not string-based, however we make a few changes to not inconvenience
our users:

  - simple_decorator() will auto-convert string arguments into a
    bytes-based decorator

  - If the given output stream is in text mode, we auto-drill down to its
    underlying .buffer to get a binary output stream.

This also allows us to do the internal line buffering in a bytearray,
which is much nicer than the single-element mutable list hack used
previously.