LogContextDemuxer: Propagate formatter from demux to context handlers
Currently, the context handlers added to the demuxer typically have no
log formatter set: the demuxer does not set any, and most context
handlers are created on the fly by logmuxed_work, which also does not
set any formatter. The only way to set a formatter is either to pass in
a custom log handler (with a formatter set), or by overriding the
default_log_handler_factory() in the logmuxed_work.Context class.

It is more desirable to set a single formatter on the demux object
itself, and have that be reused by all the context handlers (unless they
opt out).

This patch teaches LogContextDemuxer to automatically copy its
.formatter to the context handlers when they do not already have a
.formatter set. This allows users to set the their chosen formatter
once, on the demux object, and have it automatically be reused in all
contexts.

For the fallback handlers, no change is made, so log records that fall
through to these will get formatted as before.