logmuxed_work.Scheduler: Allow specifying demux log formatter
This adds the log_formatter argument to the logmuxed_work.Scheduler
contructor, which allows users to pass a logging.Formatter object to be
automatically used by all log statements that are dispatched through the
scheduler's LogContextDemuxer instance.

Also allow simply passing log_formatter=True, in which case, the
first formatter associated with the root logger will be reused.