logmux.simple_decorator(): Use '{}' to separate prefix and suffix
Higher-level layers, like LogMux.watched_fifo() and further take a
single 'decorator' argument, and pass it via simple_decorator() when it's
not already callable. As a consequence, it's much easier to create a
decorator with a prefix and no suffix (just pass a string), than to
explicitly call simple_decorator() in order to build a decorator with
both a prefix and a suffix.

Fix this, by dropping the suffix argument and turning the prefix
argument into a more flexible pattern argument. The pattern uses a
single '{}' placeholder to separate the prefix from the suffix, and
allows all but the most complex callers to simply pass a string to
define their decorator. Note that the expansion of the '{}' format
does NOT happen when the decorator is actually _used_; it is instead
evaluated immediately when calling simple_decorator() to deduce the
prefix and suffix. Hence it does not support any more fancy string
formatting operations than simply using the '{}' placeholder exactly
once.