external_work: Make sure transport is closed on subprocess termination
When terminating subprocess as part of a controlled crash^W^Wshutdown,
we've observed the following error printed after asyncio.run() is done:

  Exception ignored in: <bound method BaseSubprocessTransport.__del__ \
      of <_UnixSubprocessTransport closed pid=17306 returncode=-15 \
      stdout=<_UnixReadPipeTransport closing fd=12 open>>>
  [snip traceback...]
  RuntimeError: Event loop is closed

This appears to be due to the underlying SubprocessTransport object
associated with one of the terminated processes not having been close()d
before its __del__ destructor is run, at which point it's too late, as
the asyncio event loop has already been closed, and the close() is
unable to do its job.

Inspection of the asyncio subprocess code reveals that one way to trigger
this close() earlier is to await proc.communicate() instead of
proc.wait(), as that will take care to close the process' .stdout and
.stderr properly.

Thinking some more about this, awaiting .communicate() instead of .wait()
surely the _only_ correct way to deal with subprocess termination when
we don't (care to) know whether or not that subprocess had pipes
connected.

Also warn in the log if this communicate() returns stdout or stderr data
that is discarded.