logmuxed_work: Prepare for limiting the amount of open FIFOs
Currently we're opening two FIFOs per job, for the entire lifetime of
the job, which means that we quickly run out of file descriptors when
the number of jobs grow.

Use the new DecoratedStream objects returned from LogMux.new_stream() to
allow for delaying the FIFO opening until we actually need it. For now
there is no substantial change in behavior, but we do decouple the point
at which we associate a job (or rather: Context object) with a logmux +
decorator from the point at which we actually open + watch the FIFOs
and enable communication through them.