logmux: Simplify implementation considerably
We've had problems with the internal Muxer service in LogMux aborting,
leaving the outer LogMux stuck, hanging on self.q.join(), and the inner
excepton silently swallowed(?).

As it happens, we don't need the Muxer service at all (it's rather an
old relic from some past non-asyncio implementation where this ran in a
separate thread).

Collapse the Muxer code into the corresponding Logmux methods and get
rid of the internal queue altogether.

As a consequence, this also enables us to convert the entire LogMux (and
DecoratedStream) API from async to sync! (Note that we're still dependent
on asyncio, to attach the FIFO reader callbacks to the event loop).