test_logmuxed_work: Demonstrate how to avoid 'Too many open files'
With the recent changes in extenrnal_work and logmuxed_work, we finally
have a strategy for avoiding OSError(24, 'Too many open files') when the
number of jobs increase:

 - Async jobs that only use the logging framework for output are now
   properly decorated _without_ having to redirect their log output via
   the logmux.

 - Async jobs that need a logmux stream should throttle the realization
   of this fd (e.g. using ctx.reserve_worker()) to prevent having too
   many jobs opening fds at once.

 - Jobs that need to open logmux streams on behalf of their own threads
   or subprocesses can now do so _within_ the .reserve_worker() context,
   thanks to the new worker ticket mechanism.

All together, this makes it straightforward for a user with an unbounded
number of jobs to still exert some some control on the number of
concurrently open file descriptors. That said, a user that fails to think
about these issues will still be able to run out of fds.

Keep the test that deliberately propvokes this issue, but also add some
other tests that verify alternative solutions on how to tackle it.