logcontext: Introduce context-sensitive log formatter
This adds the logcontext.Formatter class, which is a logging.Formatter
that decorates each log message based on the current context (the async
task or non-main thread in which the log statement was made).

It is a much more lightweight alternative to the LogContextDemuxer, and
is intended to replace it, as we no longer want to redirect log messages
to different (logmux) streams.