external_work: Introduce worker "tickets" to reserve workers for longer
Until now, the Context object has reserved a worker for the duration of
.call_in_thread() or .subprocess(). This does not allow the user to do
anything (e.g. prepare file descriptors) after the worker has been
reserved, but _before_ the thread/subprocess is started. It also does
not allow for several threads/subprocesses to be executed in sequence
while holding one single worker reserved across all of them.

Fix these issues by expanding Context.reserve_worker() to issue worker
"tickets". A ticket can be seen as a license to use one worker slot
until the ticket is released (by exiting the .reserve_worker() context).
The ticket (which is yielded into the .reserve_worker() context) can be
passed to .call_in_thread() or .(run_in_)subprocess(), which will then
reuse that ticket instead of reserving their own.

In short, this allows users that want more control to do things like:

    with ctx.reserve_worker() as myticket:
        # setup some stuff needed by thread function
        ctx.call_in_thread(func, *args, ticket=myticket)
        # ...still have this one worker reserved
        ctx.run_in_subprocess(argv, ticket=myticket)
    # worker released/ticket expires when exiting context

Use this in the logmuxed_work.subprocess() override to ensure that the
.stdout/.stderr streams are only opened _after_ we have managed to
reserve a worker for the subprocess.

Add logic and tests to make sure we cannot use a single ticket to do work
in more than one worker at a time, as well as making sure an expired
ticket cannot be used to schedule work.