decorated_work: Add API for printing/following regular files
This adds the Context.follow_file() context manager, which is similar to
Context.stdout, except that instead of opening a FIFO to echo the data
there, we open a regular file instead, and echo everything in that file
until context exit.

This can be used to simply dump a pre-written file to stdout (or stderr,
when use_stderr=True is passed), but it's probably more useful when some
process is redirecting its output to a file, and we can follow that file
to echo the same output to stdout (or stderr).