stream_mux: Add simple support for following regular files
Until now, the StreamMux is typically used to follow the output of a
subprocess: We create a FIFO, and have the subprocess write its output
into that, and StreamMux will then read that FIFO, decorate the lines
read, and write the result to the common output. However, sometimes
you'd rather "tee" the subprocess output into a file as well as through
the StreamMux. This could not easily be done, as the StreamMux watch
mechanism does not work on regular files (their file descriptors are not
select()-able).

Additionally, when watching a FIFO whose write end is running in the same
thread, it is _very_ easy to create deadlocks, where a write into the
FIFO will block (waiting for the FIFO to be read by StreamMux), and the
block is exactly what stops StreamMux from reading.

Add functionality to follow a regular file: This will open the file and
periodically read to the end of the file, passing the lines through the
decorator and to the common output.