Rename logmuxed_work to decorated_work
With the logmux name gone, we need a new name for this module as well.
Compared to external_work, this module is really about adding the
ability to decorate the stdout/stderr/logs of jobs. The redirection
offered by StreamMux is not an end in itself, but rather the means by
which we achieve decoration. Hence, decorated_work is a better
description of what this module is all about.