decorated_work: Fix use outside of .decoration() context
When we're outside the .decoration() context, ctx.stdout and .stderr is
None. Calling ctx.subprocess() in this case would fail as it tries to
enter the ctx.stdout/stderr context. Fix that, by skipping the
redirection when there are not stdout/stderr streams set up.

Also refactor the decorated_work tests to allow verifying correct
behavior when outputting out/err/log lines from outside the
.decoration() context.