decorated_work: Allow users to cat/tail regular files with decoration
Add .cat_file() to DecoratedStream, to allow decorated_work users to cat
or tail files directly to ctx.stdout or ctx.stderr.

This initial implementation is sync, although we keep the API async, as
we probably want to rewrite this to not block when reading the
(potentially large) file from disk, or writing it into the mux.

Another optimization to be done is a proper tail implementation that
reads blocks from the end of the file until the requested number of
lines is found.

Note that the decorated_work and stream_mux modules need some heavy
rework, as the current API and factoring does not make much sense.